# Optimize moe