--- a/lmdeploy/pytorch/engine/model_agent.py
+++ b/lmdeploy/pytorch/engine/model_agent.py
@@ -509,11 +509,13 @@ class BaseModelAgent(AutoModelAgent):
                      adapters: Dict[str, str] = None,
                      trust_remote_code: bool = True):
         """build patched model."""
+        device = 'cuda'
         with LoadNoInit():
             hf_model = AutoModelForCausalLM.from_pretrained(
                 model_path,
                 torch_dtype=torch_dtype,
                 trust_remote_code=trust_remote_code,
+                device_map=device,
                 **self.model_config.init_kwargs)
             hf_model.eval()
             hf_model.config.use_cache = True
@@ -526,7 +528,6 @@ class BaseModelAgent(AutoModelAgent):
         if adapters:
             _unparam_lora_weight(patched_model)
 
-        patched_model = patched_model.cuda()
         return patched_model
 
     def paging_adapters(self, weight_maps: List[AdapterWeightMap]):