--- a/lmdeploy/pytorch/kernels/__init__.py
+++ b/lmdeploy/pytorch/kernels/__init__.py
@@ -2,6 +2,7 @@
 from .alibi_pagedattention import alibi_paged_attention_fwd
 from .apply_rotary_pos_emb import apply_rotary_pos_emb
 from .fill_kv_cache import fill_kv_cache
+from .fused_moe import fused_moe
 from .fused_rotary_emb import fused_rotary_emb
 from .multinomial_sampling import multinomial_sampling
 from .pagedattention import paged_attention_fwd
@@ -9,7 +10,7 @@ from .rerope_attention import rerope_attention_fwd
 from .rms_norm import rms_norm
 
 __all__ = [
-    'apply_rotary_pos_emb', 'fused_rotary_emb', 'paged_attention_fwd',
-    'alibi_paged_attention_fwd', 'fill_kv_cache', 'multinomial_sampling',
-    'rms_norm', 'rerope_attention_fwd'
+    'apply_rotary_pos_emb', 'fused_moe', 'fused_rotary_emb',
+    'paged_attention_fwd', 'alibi_paged_attention_fwd', 'fill_kv_cache',
+    'multinomial_sampling', 'rms_norm', 'rerope_attention_fwd'
 ]