deepseek-moe-16b-chat

```
concurrency: 256
elapsed_time: 250.754s

first token latency(s)(min, max, ave): 1.940, 13.106, 3.160
per-token latency(s) percentile(50, 75, 95, 99): [0.052, 0.067, 0.195, 0.229]

number of prompt tokens: 721793
number of completion tokens: 668265
token throughput (completion token): 2665.019 token/s
token throughput (prompt + completion token): 5543.506 token/s
RPS (request per second): 11.964 req/s
RPM (request per minute): 717.834 req/min
```

Qwen1.5-MoE-A2.7B-Chat TP=2

```
concurrency: 256
elapsed_time: 224.461s

first token latency(s)(min, max, ave): 1.834, 14.229, 3.098
per-token latency(s) percentile(50, 75, 95, 99): [0.057, 0.06, 0.13, 0.154]

number of prompt tokens: 680073
number of completion tokens: 620970
token throughput (completion token): 2766.490 token/s
token throughput (prompt + completion token): 5796.291 token/s
RPS (request per second): 13.365 req/s
RPM (request per minute): 801.920 req/min
```

Mixtral 8x7b TP=2

```
concurrency: 256
elapsed_time: 396.564s

first token latency(s)(min, max, ave): 2.646, 35.353, 6.054
per-token latency(s) percentile(50, 75, 95, 99): [0.07, 0.074, 0.308, 0.373]

number of prompt tokens: 741804
number of completion tokens: 712850
token throughput (completion token): 1797.565 token/s
token throughput (prompt + completion token): 3668.142 token/s
RPS (request per second): 7.565 req/s
RPM (request per minute): 453.899 req/min

```