@zhulinJulia24 2024-05-08
**Qwen1.5-MoE-A2.7B-Chat**



**main**

concurrency: 256
elapsed_time: 4845.707s

first_token latency(min, max, ave): 0.004s, 3.831s, 1.272s

number of prompt tokens: 1148381
number of completion tokens: 1016401
token throughput (completion token): 209.753 token/s
token throughput (prompt + completion token): 446.742 token/s
RPS (request per second): 1.032 req/s
RPM (request per minute): 61.910 req/min

**this pr**
concurrency: 256
elapsed_time: 1865.420s

first_token latency(min, max, ave): 0.006s, 2.110s, 0.436s

number of prompt tokens: 1148381
number of completion tokens: 1016401
token throughput (completion token): 544.864 token/s
token throughput (prompt + completion token): 1160.480 token/s
RPS (request per second): 2.680 req/s
RPM (request per minute): 160.822 req/min


**vllm is on way**