@zhulinJulia24 2024-05-09
**mistralai/Mistral-7B-Instruct-v0.1  单卡**


== Serving Benchmark Result ==
Successful requests:                     1000      
Benchmark duration (s):                  86.02     
Total input tokens:                      241080    
Total generated tokens:                  173935    
Request throughput (req/s):              11.62     
Input token throughput (tok/s):          2802.46   
Output token throughput (tok/s):         2021.93   
--Time to First Token--
Mean TTFT (ms):                          16705.49  
Median TTFT (ms):                        6305.76   
P99 TTFT (ms):                           58420.31  
--Time per Output Token (excl. 1st token)--
Mean TPOT (ms):                          113.28    
Median TPOT (ms):                        108.95    
P99 TPOT (ms):                           385.06