@zhulinJulia24 2024-05-18
<body>
<!--StartFragment-->

dataset | version | metric | mode | internlm2-chat-7b-pytorch | internlm2-chat-20b-pytorch | llama-2-7b-chat-pytorch | qwen1.5-7b-chat-pytorch | qwen1.5-moe-2.7b-chat-pytorch | llama-3-8b-instruct-pytorch
-- | -- | -- | -- | -- | -- | -- | -- | -- | --
ceval | - | naive_average | gen | 58.2 | 63.45 | 28.51 | 70.67 | 77.24 | 50.63
mmlu | - | naive_average | gen | 58.41 | 60.96 | 35.63 | 61.47 | 60.87 | 54.5
WiC | d06864 | accuracy | gen | 58.62 | 60.19 | 0 | 63.01 | 60.03 | 21.32
WSC | 7902a7 | accuracy | gen | 56.73 | 50 | 0 | 39.42 | 40.38 | 30.77
triviaqa | 2121ce | score | gen | 56.82 | 63.87 | 56.12 | 44.63 | 54.59 | 63.83
gsm8k | 1d7fe4 | accuracy | gen | 33.36 | 53.6 | 14.03 | 5.61 | 22.21 | 25.78
race-middle | 9a54b6 | accuracy | gen | 76.32 | 86.84 | 58.57 | 87.6 | 82.31 | 86.35
race-high | 9a54b6 | accuracy | gen | 75.19 | 83.7 | 51.17 | 82.7 | 79.3 | 80.93


<!--EndFragment-->
</body>

</html>