@lvhan028 2024-05-19
Hi, @zhulinJulia24, please help double check the inference speed