@zhulinJulia24 2024-05-20
> **deepseek-moe-16b-chat**
> 
> **this pr**
> 
> ```
> CUDA_VISIBLE_DEVICES=6 lmdeploy serve api_server /nvme/qa_test_models/deepseek-ai/deepseek-moe-16b-chat --server-port 24333
> python benchmark/profile_restful_api.py localhost:24333 /nvme/qa_test_models/deepseek-ai/deepseek-moe-16b-chat ../ShareGPT_V3_unfiltered_cleaned_split.json --concurrency 256 --stream-output True --num-prompts 1000
> ```
> 
> in process
> 
> **vllm 0.4.0**
> 
> ```
> CUDA_VISIBLE_DEVICES=6 python3 -m vllm.entrypoints.openai.api_server --model /nvme/qa_test_models/deepseek-ai/deepseek-moe-16b-chat --trust-remote-code
> python3 benchmarks/benchmark_serving.py --model /nvme/qa_test_models/deepseek-ai/deepseek-moe-16b-chat --dataset-path ./ShareGPT_V3_unfiltered_cleaned_split.json
> ```
> 
> Successful requests: 1000 
Benchmark duration (s): 83.97 
Total input tokens: 236158 
Total generated tokens: 158473
 Request throughput (req/s): 11.91 
Input token throughput (tok/s): 2812.46 
Output token throughput (tok/s): 1887.29 
--Time to First Token
-- Mean TTFT (ms): 16670.32 Median 
TTFT (ms): 5263.92 P99 
TTFT (ms): 57460.03 
--Time per Output Token (excl. 1st token)
-- Mean TPOT (ms): 120.28 
Median TPOT (ms): 115.89 
P99 TPOT (ms): 329.52


newest code, more quicker than before, but stil slower compare to vllm

concurrency: 256
elapsed_time: 151.330s

first_token latency(min, max, ave): 1.439s, 55.812s, 14.035s

number of prompt tokens: 232957
number of completion tokens: 224176
token throughput (completion token): 1481.376 token/s
token throughput (prompt + completion token): 3020.777 token/s
RPS (request per second): 6.608 req/s
RPM (request per minute): 396.486 req/min